feat #65 Hub 모듈 Cors 설정, 엔티티 수정, RabbitMQ 설정 수정

+++ tracky-emulator/src/main/java/kernel360/trackyemulator/infrastructure/dto/CarOnOffRequest.java
@@ -22,15 +22,15 @@ public class CarOnOffRequest {
 	private LocalDateTime offTime;
 
 	private String gcd;
-	private long lat;
-	private long lon;
+	private int lat;
+	private int lon;
 	private int ang;
 	private int spd;
 	private double sum;
 
 	private CarOnOffRequest(String mdn, String tid, String mid, String pv, String did,
 		LocalDateTime onTime, LocalDateTime offTime,
-		String gcd, long lat, long lon, int ang, int spd, double sum) {
+		String gcd, int lat, int lon, int ang, int spd, double sum) {
 		this.mdn = mdn;
 		this.tid = tid;
 		this.mid = mid;
